+++ database/create_database.py
@@ -1,5 +1,5 @@
 import mysql.connector
-from config import USER, PASSWORD, HOST
+from database.config import USER, PASSWORD, HOST
 
 """
 This is a script you run at the beginning to create the SQL Database at the beginning
